--- actor/base/pkg/types/detect.go
@@ -1,11 +1,0 @@
-package types
-
-type DetectFile struct {
-	Branches map[string]string `json:"branches"`
-	Tags     map[string]string `json:"tags"`
-}
-
-const (
-	MapKeyLatestTagHash = "latest/hash"
-	MapKeyLatestTagName = "latest/name"
-)
